fix: 4 issues from deployment report

+++ app/controllers/api/chanels/messages_controller.rb
@@ -1,4 +1,4 @@
-class Api::Chanels::MessagesController < Api::BaseController
+class Api::Channels::MessagesController < Api::BaseController
   before_action :doorkeeper_authorize!, only: %i[index destroy]
 
   def index
@@ -13,6 +13,7 @@ class Api::Chanels::MessagesController < Api::BaseController
     raise ActiveRecord::RecordNotFound if @message.blank?
 
     authorize @message, policy_class: Api::Chanels::MessagesPolicy
+    authorize @message, policy_class: Api::Channels::MessagesPolicy
 
     if @message.destroy
       head :ok, message: I18n.t('common.200')
